// src/app/interfaces/irequest.ts
export interface IRequest {
    requestNumber: string; 
    Ttype: string; 
    signal: string; 
    bookNumber: string; 
    requesterName: string; 
    requesterCivilNo: string; 
    companyName: string; 
    kuwaitIn: string; 
    centeralNumber: string; 
    orgCivilNo: string; 
    MtNo: string; 
    TNo: string; 
    EN: string;
    address: string;
    createdOn: string;
}